refactor(automation): merge planner_claude into planner
Merge planner_claude.py (210 lines) into planner.py. The module has a single consumer and its 5 functions are private implementation details of the planning domain. Reduces module count and eliminates unnecessary indirection while maintaining the same imports and abstractions.

Closes #1444

Implemented-By: claude-opus-4-8
