+++ hephaestus/automation/planner.py
@@ -11,7 +11,9 @@ from __future__ import annotations
 
 import argparse
 import logging
+import subprocess
 import threading
+import time
 from concurrent.futures import Future, ThreadPoolExecutor, as_completed
 from pathlib import Path
 from typing import Any
@@ -19,6 +21,7 @@ from typing import Any
 from hephaestus.agents.runtime import (
     direct_agent_model,
     resolve_agent,
+    run_agent_text,
     uses_direct_agent_runner,
 )
 from hephaestus.cli.utils import (
@@ -29,6 +32,7 @@ from hephaestus.cli.utils import (
     emit_json_status,
 )
 from hephaestus.constants import AUTOMATION_LOG_FORMAT, LOG_DATEFMT
+from hephaestus.github.rate_limit import wait_until
 
 from ._review_utils import (
     build_automation_parser,
@@ -37,20 +41,20 @@ from ._review_utils import (
     find_pr_for_issue,
 )
 from .advise_runner import advise_skipped, ensure_mnemosyne, run_advise
+from .claude_invoke import detect_server_overload, invoke_claude_with_session, scan_quota_reset
 from .claude_models import advise_model, codex_advise_model
 from .claude_timeouts import (
     DEFAULT_AGENT_TIMEOUT,
     DEFAULT_GIT_MESSAGE_AGENT_TIMEOUT,
     planner_claude_timeout,
 )
-from .git_utils import issue_ref
+from .git_utils import get_repo_root, get_repo_slug, issue_ref
 from .github_api import (
     GitHubRateLimitError,
     gh_issue_upsert_comment,
     gh_list_open_issues,
 )
 from .models import PLAN_COMMENT_MARKER, PlannerOptions, PlanResult
-from .planner_claude import PlannerClaudeRunner
 from .planner_review_loop import MAX_REVIEW_ITERATIONS, PlanReviewLoop
 from .planner_state import PlannerStateManager
 from .prompts import get_advise_prompt_builder
@@ -62,6 +66,206 @@ from .work_report import work_report_context
 __all__ = ["MAX_REVIEW_ITERATIONS", "Planner", "main"]
 
 logger = logging.getLogger(__name__)
+
+# Base delay (seconds) for the exponential backoff applied to transient
+# server-overload (529 / 5xx) retries. The Nth retry (counting down from the
+# default ``max_retries=3``) waits ``_OVERLOAD_BACKOFF_BASE_S * 2 ** (3 -
+# max_retries)`` seconds, i.e. 5s, 10s, 20s. Unlike a 429 quota cap there is no
+# reset epoch to wait on, so a bounded exponential backoff is the correct
+# policy (#1374).
+_OVERLOAD_BACKOFF_BASE_S = 5.0
+# Default retry budget the backoff schedule is anchored to; used only to size
+# the per-attempt delay so it grows monotonically as retries are consumed.
+_OVERLOAD_BACKOFF_ANCHOR_RETRIES = 3
+
+
+class PlannerClaudeRunner:
+    """Run Claude or Codex on behalf of the planner.
+
+    Holds a reference to :class:`PlannerOptions` so it can pick the right
+    backend (Claude vs Codex) and forward the optional system-prompt file.
+    """
+
+    def __init__(self, options: PlannerOptions) -> None:
+        """Bind to the planner's options.
+
+        Args:
+            options: The shared :class:`PlannerOptions` instance.
+
+        """
+        self.options = options
+
+    def call_claude(
+        self,
+        prompt: str,
+        *,
+        model: str,
+        agent: str,
+        issue_number: int | str,
+        max_retries: int = 3,
+        timeout: int | None = None,
+        extra_args: list[str] | None = None,
+    ) -> str:
+        """Call Claude CLI on a deterministic session with rate-limit retry.
+
+        The session UUID is derived from ``(repo, issue_number, agent,
+        trunk_githash)`` via :func:`session_naming.session_uuid`. First call
+        for a tuple creates the session; every later call resumes it. Cross-
+        agent independence (planner vs reviewer) is preserved because the
+        ``agent`` string is part of the hash.
+
+        Args:
+            prompt: The prompt to send to Claude.
+            model: Claude model ID for ``--model`` (caller picks per phase).
+            agent: One of the ``AGENT_*`` constants from
+                :mod:`hephaestus.automation.session_naming`. Different agents
+                map to different session IDs.
+            issue_number: GitHub issue number; participates in the session ID.
+            max_retries: Maximum retry attempts for rate limits.
+            timeout: Subprocess timeout in seconds.
+            extra_args: Additional CLI arguments.
+
+        Returns:
+            Claude's response text (stdout, stripped).
+
+        Raises:
+            RuntimeError: If Claude call fails.
+
+        """
+        timeout_s = planner_claude_timeout() if timeout is None else timeout
+        if uses_direct_agent_runner(self.options.agent):
+            return self.call_direct_agent(
+                prompt,
+                model=direct_agent_model(
+                    self.options.agent,
+                    "HEPH_PLANNER_MODEL",
+                    codex_default=model,
+                ),
+                max_retries=max_retries,
+                timeout=timeout_s,
+            )
+
+        repo_root = get_repo_root()
+        repo = get_repo_slug(repo_root)
+
+        try:
+            stdout, _ = invoke_claude_with_session(
+                repo=repo,
+                issue=issue_number,
+                agent=agent,
+                prompt=prompt,
+                model=model,
+                cwd=repo_root,
+                timeout=timeout_s,
+                system_prompt_file=self.options.system_prompt_file,
+                allowed_tools="Read,Glob,Grep,Bash",
+                extra_args=extra_args,
+            )
+            response = stdout.strip()
+            if not response:
+                raise RuntimeError("Claude returned empty response")
+            return response
+
+        except subprocess.CalledProcessError as e:
+            stderr = e.stderr or ""
+            stdout = e.stdout or ""
+
+            # Rate-limit messages may appear in either stream. The Claude CLI
+            # in particular returns its 429 ("You're out of extra usage ·
+            # resets ...") inside the stdout JSON payload as the ``result``
+            # field of an ``is_error: true`` response, not in stderr.
+            reset_epoch = scan_quota_reset(stderr, stdout)
+            if reset_epoch is not None and max_retries > 0:
+                if reset_epoch > 0:
+                    wait_until(reset_epoch)
+                else:
+                    time.sleep(5)
+                return self.call_claude(
+                    prompt,
+                    model=model,
+                    agent=agent,
+                    issue_number=issue_number,
+                    max_retries=max_retries - 1,
+                    timeout=timeout_s,
+                    extra_args=extra_args,
+                )
+
+            # A transient server-overload (529 Overloaded / generic 5xx) carries
+            # no reset epoch, so scan_quota_reset misses it. It used to fall
+            # straight through to the fatal raise below, surfacing a recoverable
+            # blip as ``phase plan FAILED rc=1`` (#1374). Retry it with bounded
+            # exponential backoff up to ``max_retries``.
+            if detect_server_overload(stderr, stdout) and max_retries > 0:
+                # Exponent grows as retries are consumed (clamped at 0 so an
+                # over-budget caller never gets a fractional/negative delay).
+                exponent = max(0, _OVERLOAD_BACKOFF_ANCHOR_RETRIES - max_retries)
+                delay = _OVERLOAD_BACKOFF_BASE_S * (2**exponent)
+                logger.warning(
+                    "Claude server overloaded (transient); retrying in %.0fs (%d retries left)",
+                    delay,
+                    max_retries,
+                )
+                time.sleep(delay)
+                return self.call_claude(
+                    prompt,
+                    model=model,
+                    agent=agent,
+                    issue_number=issue_number,
+                    max_retries=max_retries - 1,
+                    timeout=timeout_s,
+                    extra_args=extra_args,
+                )
+
+            detail = stderr or stdout or "(no output)"
+            raise RuntimeError(f"Claude failed: {detail}") from e
+
+        except subprocess.TimeoutExpired as e:
+            raise RuntimeError(f"Claude timed out after {timeout_s}s") from e
+
+    def call_direct_agent(
+        self,
+        prompt: str,
+        *,
+        model: str,
+        max_retries: int = 3,
+        timeout: int | None = None,
+        sandbox: str = "workspace-write",
+    ) -> str:
+        """Call a non-Claude direct agent with retry logic for rate limits."""
+        agent = self.options.agent
+        timeout_s = planner_claude_timeout() if timeout is None else timeout
+        try:
+            result = run_agent_text(
+                agent=agent,
+                prompt=prompt,
+                cwd=get_repo_root(),
+                timeout=timeout_s,
+                model=model,
+                sandbox=sandbox,
+            )
+        except subprocess.CalledProcessError as e:
+            stderr = e.stderr or ""
+            stdout = e.stdout or ""
+            reset_epoch = scan_quota_reset(stderr, stdout)
+            if reset_epoch is not None and reset_epoch > 0 and max_retries > 0:
+                logger.warning("%s usage cap hit; waiting for reset", agent)
+                wait_until(reset_epoch)
+                return self.call_direct_agent(
+                    prompt,
+                    model=model,
+                    max_retries=max_retries - 1,
+                    timeout=timeout_s,
+                    sandbox=sandbox,
+                )
+            detail = stderr or stdout or str(e)
+            raise RuntimeError(f"{agent} failed: {detail}") from e
+        except subprocess.TimeoutExpired as e:
+            raise RuntimeError(f"{agent} timed out after {timeout_s}s") from e
+
+        response = (result.stdout or "").strip()
+        if not response:
+            raise RuntimeError(f"{agent} returned empty response")
+        return response
 
 
 class Planner:
