+++ tests/unit/automation/test_planner.py
@@ -54,10 +54,10 @@ class TestCallClaude:
 
         stack = ExitStack()
         stack.enter_context(
-            patch("hephaestus.automation.planner_claude.get_repo_root", return_value=Path("/repo"))
+            patch("hephaestus.automation.planner.get_repo_root", return_value=Path("/repo"))
         )
         stack.enter_context(
-            patch("hephaestus.automation.planner_claude.get_repo_slug", return_value="TestRepo")
+            patch("hephaestus.automation.planner.get_repo_slug", return_value="TestRepo")
         )
         return stack
 
@@ -159,7 +159,7 @@ class TestCallClaude:
         with (
             self._patch_repo(),
             patch("hephaestus.automation.claude_invoke.subprocess.run") as mock_run,
-            patch("hephaestus.automation.planner_claude.scan_quota_reset") as mock_scan,
+            patch("hephaestus.automation.planner.scan_quota_reset") as mock_scan,
         ):
             mock_run.side_effect = [
                 subprocess.CalledProcessError(1, "claude", stderr="rate limit exceeded"),
@@ -188,7 +188,7 @@ class TestCallClaude:
         with (
             self._patch_repo(),
             patch("hephaestus.automation.claude_invoke.subprocess.run") as mock_run,
-            patch("hephaestus.automation.planner_claude.time.sleep") as mock_sleep,
+            patch("hephaestus.automation.planner.time.sleep") as mock_sleep,
         ):
             mock_run.side_effect = [
                 subprocess.CalledProcessError(
@@ -216,7 +216,7 @@ class TestCallClaude:
         with (
             self._patch_repo(),
             patch("hephaestus.automation.claude_invoke.subprocess.run") as mock_run,
-            patch("hephaestus.automation.planner_claude.time.sleep"),
+            patch("hephaestus.automation.planner.time.sleep"),
         ):
             mock_run.side_effect = subprocess.CalledProcessError(
                 1, "claude", stderr="API Error: 529 Overloaded"
@@ -237,7 +237,7 @@ class TestCallClaude:
         with (
             self._patch_repo(),
             patch("hephaestus.automation.claude_invoke.subprocess.run") as mock_run,
-            patch("hephaestus.automation.planner_claude.time.sleep") as mock_sleep,
+            patch("hephaestus.automation.planner.time.sleep") as mock_sleep,
         ):
             mock_run.side_effect = subprocess.CalledProcessError(
                 1, "claude", stderr="API Error: 400 Bad Request"
@@ -262,7 +262,7 @@ class TestCallClaude:
         with (
             self._patch_repo(),
             patch("hephaestus.automation.claude_invoke.subprocess.run") as mock_run,
-            patch("hephaestus.automation.planner_claude.wait_until") as mock_wait,
+            patch("hephaestus.automation.planner.wait_until") as mock_wait,
         ):
             mock_run.side_effect = [
                 subprocess.CalledProcessError(1, "claude", output=usage_json, stderr=""),
