+++ tests/unit/automation/test_provider_neutral_direct_dispatch.py
@@ -12,7 +12,6 @@ REPO_ROOT = Path(__file__).resolve().parents[3]
 PROVIDER_NEUTRAL_FILES = [
     "hephaestus/automation/agent_stage.py",
     "hephaestus/automation/planner.py",
-    "hephaestus/automation/planner_claude.py",
     "hephaestus/automation/plan_reviewer.py",
     "hephaestus/automation/_implement_phase.py",
     "hephaestus/automation/implementer_phase_runner.py",
